requirements.txt
asgiref==3.6.0
Django==4.2.1
djangorestframework==3.14.0
environ==1.0
Pillow==9.5.0
psycopg2-binary==2.9.5
pydantic==1.10.7
PyJWT==2.6.0
pytz==2023.3
PyYAML==6.0
sqlparse==0.4.4
typing_extensions==4.5.0
tzdata==2023.3
uuid==1.30
